Move to systemtests as utils package
