--- utils/data_archive/tests/__init__.py
@@ -1,6 +1,0 @@
-# ############################################################################### #
-# Autoreduction Repository : https://github.com/ISISScientificComputing/autoreduce
-#
-# Copyright &copy; 2020 ISIS Rutherford Appleton Laboratory UKRI
-# SPDX - License - Identifier: GPL-3.0-or-later
-# ############################################################################### #
